Changes for 0.9.9
* The build tools can now generate bundles with relative paths. Fixes #19
* Fixed minor typos in the generated template files closed #17
* Renamed generators to sc_generators.  This avoids a conflict with Rails 2.1.
* [FIX] Applied bug fixes suggested by typo_pl to make the build tool run on windows.

+++ lib/sproutcore/merb/bundle_controller.rb
@@ -93,10 +93,14 @@ module SproutCore
         ret = File.open(build_path)
         
         
-        # In development mode only, immediately delete built composite resources.  We want 
-        # each request to come directly to us.
+        # In development mode only, immediately delete built composite 
+        # resources.  We want each request to come directly to us.
         if (current_bundle.build_mode == :development) && (!entry.cacheable?)
-          FileUtils.rm(build_path)
+          
+          # Deleting composite resources will not work in windows because it
+          # does not like to have files you just open deleted. (Its OK on
+          # windows)
+          FileUtils.rm(build_path) if (RUBY_PLATFORM !~ /mswin32/)
         end
         
         return ret
